requirements.txt
Flask==2.3.0
Pillow==9.5.0
requests==2.31.0
Werkzeug==2.3.0
gunicorn==21.2.0
